requirements.txt
anyio==3.3.2
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
colorama==0.4.4
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.6
djangorestframework==3.12.4
entrypoints==0.3
idna==3.2
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==4.0.1
jupyter-client==7.0.5
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
prometheus-client==0.11.0
prompt-toolkit==3.0.20
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
# pywin32==301
# pywinpty==1.1.4
pyzmq==22.3.0
requests==2.26.0
requests-unixsocket==0.2.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.0
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
djangorestframework==3.12.4
django-crispy-forms